refactor(aisections): move overlays into feature folder
Both AI Sections overlays now live next to the rest of the feature's code
(`AISectionsOverview`, `CornerHandles`, `AddSectionDialog`, `shared/...`)
instead of inside `schema-editor/viewports/`. Tests follow via `git mv`
into the existing `__tests__/` subfolder.

Only consumer is `src/lib/editor/bindings.ts` — import paths updated.
Internal `./WorldViewport`/`./WorldViewport.types` refs in the overlays
are now alias-pathed since they reach out of the feature.

Refs #48, #49, #59 (partial — WorkspaceHierarchy is deferred).
